cope with moving detApi to apiConfig
